+++ example/lib/app/modules/home/home_controller.dart
@@ -1,5 +1,9 @@
 // ignore_for_file: unnecessary_overrides
 
+import 'dart:io';
+
+import 'package:file_picker/file_picker.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:whatsapp_bot_flutter/whatsapp_bot_flutter.dart';
@@ -14,12 +18,21 @@ class HomeController extends GetxController {
   var countryCode = TextEditingController();
   var phoneNumber = TextEditingController();
   var formKey = GlobalKey<FormState>();
+  Rx<ConnectionEvent?> connectionEvent = Rxn<ConnectionEvent>();
 
   @override
   void onInit() {
     countryCode.text = "91";
     phoneNumber.text = "";
     message.text = "Testing Whatsapp Bot";
+
+    connectionEvent.bindStream(WhatsappBotFlutter.connectionEventStream);
+
+    WhatsappBotFlutter.messageEvents.listen((Message message) {
+      if (!(message.id?.fromMe ?? true)) {
+        Get.log(message.toJson().toString());
+      }
+    });
     super.onInit();
   }
 
@@ -47,17 +60,49 @@ class HomeController extends GetxController {
     );
   }
 
+  void disconnect() async {
+    await WhatsappBotFlutter.disconnect();
+    connected.value = false;
+  }
+
+  void testMethod() async {
+    try {
+      sendFileMessage();
+    } catch (e) {
+      Get.log(e.toString());
+    }
+  }
+
   void sendMessage() async {
     if (!formKey.currentState!.validate()) return;
     try {
-      await WhatsappBotFlutter.sendMessage(
+      await WhatsappBotFlutter.sendTextMessage(
         countryCode: countryCode.text,
         phone: phoneNumber.text,
         message: message.text,
-        progress: (int prg) {
-          progress.value = prg;
-          Get.log(prg.toString());
-        },
+      );
+    } catch (e) {
+      Get.log("Error : $e");
+    }
+  }
+
+  void sendFileMessage() async {
+    if (!formKey.currentState!.validate()) return;
+    try {
+      FilePickerResult? result = await FilePicker.platform.pickFiles(
+        type: FileType.image,
+      );
+      String? path = result?.files.first.path;
+      if (path == null) return;
+      File file = File(path);
+      Uint8List imageBytes = Uint8List.fromList(file.readAsBytesSync());
+
+      await WhatsappBotFlutter.sendFileMessage(
+        countryCode: countryCode.text,
+        phone: phoneNumber.text,
+        fileBytes: imageBytes,
+        caption: message.text,
+        fileType: WhatsappFileType.image,
       );
     } catch (e) {
       Get.log("Error : $e");