+++ lib/core.dart
@@ -1,5 +1,0 @@
-/*
-We believe, the class name must be unique. 
-If there is a conflicting class name in this file,
-it means you have to rename it to something more unique.
-*/