+++ lib/src/model/connection_event.dart
@@ -1,0 +1,9 @@
+enum ConnectionEvent {
+  authenticated,
+  logout,
+  authCodeChange,
+  mainLoaded,
+  mainReady,
+  qrcodeIdle,
+  requireAuth
+}