+++ lib/src/model/whatsapp_exception.dart
@@ -1,0 +1,16 @@
+enum WhatsappExceptionType {
+  unknown,
+  connectionFailed,
+  unAuthorized,
+  inValidContact,
+  clientNotConnected
+}
+
+class WhatsappException implements Exception {
+  late String message;
+  late WhatsappExceptionType exceptionType;
+  WhatsappException({
+    this.message = "Something went wrong",
+    this.exceptionType = WhatsappExceptionType.unknown,
+  });
+}