+++ lib/src/model/whatsapp_file_type.dart
@@ -1,0 +1,3 @@
+enum WhatsappFileType { document, image, audio, video }
+
+