+++ lib/src/whatsapp_bot_flutter.dart
@@ -1,6 +1,12 @@
 // ignore_for_file: avoid_print
 
-import 'helper/puppeteer_service.dart';
+import 'package:flutter/foundation.dart';
+import 'package:whatsapp_bot_flutter/src/model/connection_event.dart';
+import 'package:whatsapp_bot_flutter/src/model/message.dart';
+import 'package:whatsapp_bot_flutter/src/model/whatsapp_file_type.dart';
+import 'package:whatsapp_bot_flutter/src/wpp/wpp_events.dart';
+
+import 'puppeteer_service.dart';
 
 class WhatsappBotFlutter {
   static final _puppeteerService = PuppeteerService();
@@ -16,7 +22,7 @@ class WhatsappBotFlutter {
     Duration? connectionTimeout,
   }) async {
     // dispose any pending tasks first
-    dispose();
+    disconnect();
     // Try to connect and login again
     await _puppeteerService.connectAndLogin(
       sessionDirectory: sessionDirectory,
@@ -29,24 +35,67 @@ class WhatsappBotFlutter {
     );
   }
 
-  // call dispose to close all resources
-  static dispose() {
+  /// call [disconnect] to dispose browse and close all resources
+  static disconnect() {
     _puppeteerService.dispose();
   }
 
   /// [sendMessage] to send messages to the given phone number
   /// listen for progress updates by `progress` callback
-  static Future<void> sendMessage({
+  static Future<void> sendTextMessage({
     required String phone,
     required String countryCode,
     required String message,
-    Function(int)? progress,
   }) async {
-    await _puppeteerService.sendMessage(
+    await _puppeteerService.sendTextMessage(
       countryCode: countryCode,
       phone: phone,
       message: message,
-      progress: progress,
     );
   }
+
+  static Future<void> sendFileMessage({
+    required String phone,
+    required String countryCode,
+    required Uint8List fileBytes,
+    required WhatsappFileType fileType,
+    String? caption,
+    String? mimetype,
+  }) async {
+    await _puppeteerService.sendFileMessage(
+      countryCode: countryCode,
+      phone: phone,
+      caption: caption,
+      fileBytes: fileBytes,
+      fileType: fileType,
+      mimetype: mimetype,
+    );
+  }
+
+  static Future<void> sendLocationMessage({
+    required String phone,
+    required String countryCode,
+    required String lat,
+    required String long,
+    String? name,
+    String? address,
+    String? url,
+  }) async {
+    await _puppeteerService.sendLocationMessage(
+      countryCode: countryCode,
+      phone: phone,
+      lat: lat,
+      long: long,
+      address: address,
+      name: name,
+      url: url,
+    );
+  }
+
+  /// [connectionEventStream] will give update of Connection Events
+  static Stream<ConnectionEvent> connectionEventStream =
+      WppEvents.connectionEventStreamController.stream;
+
+  static Stream<Message> messageEvents =
+      WppEvents.messageEventStreamController.stream;
 }