+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 ## 0.0.2
 
 - minor fixes
+
+## 0.0.3
+
+- major refactoring
+- added option to share files
+- added option to share location
+- added option to get messages stream